secure application using OAuth2
the application will now prevent access exept from user authenticated
via GitHub. For this the application was registered with GitHub. The
ClientID and ClientSecret of the application are checked in. This is no
problem as GitHub is configured to send the user back to localhost:8080
so no other site should be capabel of stealing the access token.